refactor: 레이스 횟수 관련 코드 setRaceLaps 메서드로 이동하고 적용
